Make a mixin to keep things clean
